Move command-related classes into the command module
