+++ src/main/java/com/chopshop166/chopshoplib/commands/CommandRobot.java
@@ -1,4 +1,4 @@
-package com.chopshop166.chopshoplib;
+package com.chopshop166.chopshoplib.commands;
 
 import java.io.IOException;
 import java.net.URL;
@@ -7,6 +7,8 @@ import java.util.function.Consumer;
 import java.util.jar.Attributes;
 import java.util.jar.Manifest;
 
+import com.chopshop166.chopshoplib.RobotMapFor;
+import com.chopshop166.chopshoplib.RobotUtils;
 import com.google.common.io.Resources;
 import com.google.common.reflect.ClassPath;
